app/requirements.txt
APScheduler==3.6.3
bcrypt==3.1.7
blinker==1.4
certifi==2020.6.20
cffi==1.14.0
click==7.1.1
cryptography==2.9.2
decorator==4.4.2
Flask==1.1.2
Flask-Mail==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.0
pgpubsub==0.0.5
psycopg2==2.8.5
pycparser==2.20
python-telegram-bot==12.8
pytz==2019.3
six==1.14.0
tornado==6.0.4
tzlocal==2.0.0
uWSGI==2.0.18
Werkzeug==1.0.1
